migration check

+++ database/migrations/2023_06_28_113636_add_at_salon_col_to_service_bookings_table.php
@@ -27,8 +27,12 @@ class AddAtSalonColToServiceBookingsTable extends Migration
     public function down()
     {
         Schema::table('service_bookings', function (Blueprint $table) {
-            $table->dropColumn('at_salon');
-            $table->dropColumn('remark');
+            if (Schema::hasColumn('service_bookings', 'at_salon')){
+                $table->dropColumn('at_salon');
+            }
+            if (Schema::hasColumn('service_bookings', 'remark')){
+                $table->dropColumn('remark');
+            }
         });
     }
 }
